Added vehicle class with options of creating a point mass and displaying its data

+++ Engineer_Thesis/Vector3D.cpp
@@ -7,7 +7,7 @@ Vector3D::Vector3D() {
 	z = 0;
 }
 
-Vector3D::Vector3D(float x, float y, float z) {
+Vector3D::Vector3D(double x, double y, double z) {
 
 	this->x = x;
 	this->y = y;
